Extract `getBaseThemeDependencies()` from `lookup()`
Just like the previous commits in this series, there isn't really any
dynamism left here (there was in the v8 branch, but no longer in v9), so
we can kill the entire `lookup()` abstraction.

Test plan: `yarn test`
